Move ValidPathInfo to its own header
