+++ src/libstore/nar-info.cc
@@ -1,5 +1,6 @@
 #include "globals.hh"
 #include "nar-info.hh"
+#include "store-api.hh"
 
 namespace nix {
